+++ modules/12-transient.el
@@ -82,6 +82,7 @@
   [["File"
     ("r" "Rename file"    denote-rename-file)
     ("k" "Add keywords"   denote-rename-file-keywords)
+    ("c" "Change silo"    my/denote-move-to-silo)
     ("d" "Delete note"    my/denote-delete-note)]
    [("q" "Quit" transient-quit-one)]])